top-requirements.txt
h5py
jupyter
keras
matplotlib
numpy
pyzmq
scikit-learn
tensorflow
